+++ sermant-plugins/sermant-register-center/config/config.yaml
@@ -1,4 +1,4 @@
 servicecomb.service:
-  address: http://127.0.0.1:30100 #注册中心地址
+  address: http://localhost:30100 #注册中心地址
   heartbeatInterval: 15 #服务实例心跳发送间隔（单位：秒）
   openMigration: false #是否开启迁移功能
