+++ sermant-plugins/sermant-route/gray-dubbo-2.7.x-service/src/main/java/com/huawei/gray/dubbo/utils/FieldPrivilegedAction.java
@@ -1,0 +1,41 @@
+/*
+ * Copyright (C) 2022-2022 Huawei Technologies Co., Ltd. All rights reserved.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.huawei.gray.dubbo.utils;
+
+import java.lang.reflect.Field;
+import java.security.PrivilegedAction;
+
+/**
+ * 权限检查
+ *
+ * @author provenceee
+ * @since 2022/2/21
+ */
+public class FieldPrivilegedAction implements PrivilegedAction<Object> {
+    private final Field field;
+
+    public FieldPrivilegedAction(Field field) {
+        super();
+        this.field = field;
+    }
+
+    @Override
+    public Object run() {
+        field.setAccessible(true);
+        return null;
+    }
+}
