+++ sermant-plugins/sermant-route/gray-feign-http-9.x-service/src/main/java/com/huawei/gray/feign/context/CurrentInstance.java
@@ -23,8 +23,6 @@ package com.huawei.gray.feign.context;
  * @since 2021-11-03
  */
 public class CurrentInstance {
-    private static CurrentInstance currentInstance = null;
-
     /**
      * 所属服务名称
      */
@@ -43,12 +41,6 @@ public class CurrentInstance {
     private CurrentInstance() {
     }
 
-    private CurrentInstance(String appName, String ip, int port) {
-        this.appName = appName;
-        this.ip = ip;
-        this.port = port;
-    }
-
     /**
      * 获取实例，仅在启动的时候调用，不存在线程安全问题
      *
@@ -57,13 +49,9 @@ public class CurrentInstance {
      * @param port 端口
      */
     public static void newInstance(String appName, String ip, int port) {
-        if (currentInstance == null) {
-            synchronized (CurrentInstance.class) {
-                if (currentInstance == null) {
-                    currentInstance = new CurrentInstance(appName, ip, port);
-                }
-            }
-        }
+        SingleTon.CURRENT_INSTANCE.setAppName(appName);
+        SingleTon.CURRENT_INSTANCE.setIp(ip);
+        SingleTon.CURRENT_INSTANCE.setPort(port);
     }
 
     /**
@@ -72,7 +60,7 @@ public class CurrentInstance {
      * @return 单例
      */
     public static CurrentInstance getInstance() {
-        return currentInstance;
+        return SingleTon.CURRENT_INSTANCE;
     }
 
     public String getAppName() {
@@ -97,5 +85,9 @@ public class CurrentInstance {
 
     public void setPort(int port) {
         this.port = port;
+    }
+
+    private static class SingleTon {
+        private static final CurrentInstance CURRENT_INSTANCE = new CurrentInstance();
     }
 }
